+++ internal/agent/f5/as3/extended_endpoint.go
@@ -14,10 +14,11 @@ import (
 // ExtendedEndpoint is an endpoint with additional fields...
 type ExtendedEndpoint struct {
 	models.Endpoint
-	Port             *ports.Port
-	ServicePorts     []int32
-	ServiceNetworkId strfmt.UUID
-	SegmentId        *int
-	ProxyProtocol    bool
-	Owned            bool
+	Port                *ports.Port
+	ServicePorts        []int32
+	ServiceNetworkId    strfmt.UUID
+	SegmentId           *int
+	ProxyProtocol       bool
+	Owned               bool
+	ConnectionMirroring bool
 }