+++ internal/agent/f5/endpoints_test.go
@@ -132,7 +132,7 @@ var PostBigIPFixture = &as3.AS3{
 								Label:      "endpoint-80-95dbe813-62f9-47f1-90ba-09f2dadcaefa",
 								Class:      "Service_L4",
 								AllowVlans: []string{"/Common/vlan-123"},
-								IRules:     []as3.Pointer{}, Mirroring: "L4",
+								IRules:     []as3.Pointer{}, Mirroring: "none",
 								PersistenceMethods: []string{}, Pool: as3.Pointer{
 									BigIP: "/Common/Shared/pool-a0a0a0a0-a0a0-4a0a-8a0a-0a0a0a0a0a0a-80"},
 								ProfileL4:  &as3.Pointer{BigIP: "/Common/cc_fastL4_noaging_profile"},